Add feature flags
The feature flags functionality was developed in Refer as a Rails engine
to make extracting it easier.

As part of that process, we decided to move the engine to its own repo
earlier than expected.

This allows us to develop it separately from the main work in Refer.

See PR #14 for the git history of these changes. Due to the nature of
extracting it here, I wasn't able to recreate the series of commits that
built this engine in this repo.

https://github.com/DFE-Digital/refer-serious-misconduct/pull/14