Fix CSS build
It seems that including this engine in a parent app makes it difficult
for esbuild to compile the sass.

Mainly the issue appears to be that it is tricky to reference the sass
stylesheet located in the gem from the parent app in order to compile it
using esbuild.

Rather than force every parent app to deal with the paths and compile
CSS for a dependency, I decided to include the compiled CSS as part of
the gem install.

The result is that we'll need to create a new release of the gem every
time the styles need to change. This could be because of a change in
this library or from a library it depends on, eg. govuk-components.

The trade-off for this appears worth it to ensure the experience of
getting the library running is smooth and doesn't require extra
configuration.