Use the parent layout
This engine currently defines its own layout template. This makes it
hard to integrate to a particular section of the parent app.

This change relies on the parent app defining a layout template named
`application`. This name is a Rails convention, so it is a fairly safe
assumption that it will be present.

The upshot is that this engine only needs to define the HTML for its
views, and this will behave like any other view in the parent app.

As part of this change, there is no need to include a view component for
rendering the flash message. This will be delegated to the parent app
now.