Enable access to parent app helpers
By default, the engine isn't able to access the parent app's helper
methods.

A common pattern we have used in GOVUK services is to generate the
appropriate service header in a helper method that contains the logic of
which navigation elements to include based on the route namespace.

Including all the parent's app helpers is a catch all solution to reduce
the element of surprise when building services.